🚧🏗️🔡 Move stuff around and update docs
